image_type_tezi.bbclass: Support autoinstall
Add variable TEZI_AUTOINSTALL to support automatic installation
of the Tezi image by Toradex Easy Installer. If the variable is
set to true ('yes', 'y', 'true', 't' or '1') the autoinstall
will be enabled in the generated image.json.

By default, if variable TEZI_AUTOINSTALL is not set, the
autoinstall option will be disabled.

For more details regarding the JSON image file and supported
properties please have a look at:
https://developer.toradex.com/software/toradex-easy-installer#image-format

Signed-off-by: Leon Anavi <leon.anavi@konsulko.com>