tezi-metadata: Automatic reboot after installation
Support an option to enable automatic reboot after completing
an installation by setting variable TEZI_AUTOREBOOT to 1. If
the variable is not set or has any other value the automatic
reboot after installation will be disabled.

This feature is useful to automatically reboot after automatic
installation through Toradex Easy Installer. It is a convenient
approach for headless devices.

Signed-off-by: Leon Anavi <leon.anavi@konsulko.com>